Moved CoordinatorBehaviorActivity to separate module.
