+++ coordinatorbehavior/src/main/AndroidManifest.xml
@@ -1,0 +1,8 @@
+<manifest package="com.kirchhoff.coordinatebehavior"
+	xmlns:android="http://schemas.android.com/apk/res/android">
+
+	<application>
+		<activity android:name=".CoordinatorBehaviorActivity" />
+	</application>
+
+</manifest>
