+++ coordinatorbehavior/src/main/java/com/kirchhoff/coordinatebehavior/CoordinatorBehaviorActivity.java
@@ -1,25 +1,16 @@
-package com.example.kirchhoff.example.ui.other;
+package com.kirchhoff.coordinatebehavior;
 
-import android.content.Context;
-import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 
-import com.example.kirchhoff.example.R;
-
 /**
  * @author Kirchhoff-
  */
 
 public class CoordinatorBehaviorActivity extends AppCompatActivity {
-
-
-    public static void start(Context context) {
-        context.startActivity(new Intent(context, CoordinatorBehaviorActivity.class));
-    }
 
 
     @Override
