+++ coordinatorbehavior/src/main/res/layout/a_coordinate_behavior.xml
